+++ toys-core/test/lookup-cases/lib-dirs/ns/.lib/hello.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Hello
+  def self.value
+    9
+  end
+end